// deploy/.staging-1.0.3.54-final/HIJRA-1.0.3.54-COMPLETE-SUITE-FINAL/SelfServiceSuite/SelfServicePortal/self-service-portal/src/types/erp.types.ts
export interface Attachment {
  id: string
  fileName: string
  fileType: string
  size: number
  progress: number
  uploadedAt: string
  description?: string
  contentBase64?: string
}